requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decouple==0.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.6
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-import-export==2.1.0
django-settings-export==1.2.1
et-xmlfile==1.0.1
feedparser==5.2.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
mysql-client==0.0.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.1.2
protobuf==3.11.3
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==2.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
xlwt==1.3.0